Enhance rake tasks before run time
This changes the enhancing of rake tasks to be done at the point of this
file being executed rather than the tasks being run.

This is needed so that these enhancements are compatible with Rails
engine applications - which wrap tasks under a namespace of app [1]. If
these enhancements are done when this file is executed these tasks
automatically map to the new namespace, whereas at runtime the tasks no
longer exists and cannot be executed.

[1]: https://github.com/rails/rails/blob/master/railties/lib/rails/tasks/engine.rake#L4-L17